+++ odoo/local-src/delivery_rounds/models/stock_move.py
@@ -62,6 +62,43 @@ class StockMove(models.Model):
         return res
 
     @api.multi
+    def action_done(self):
+        """ Trigger re-reserve on pickings """
+        res = super(StockMove, self).action_done()
+        received = self.filtered(lambda m: (
+            m.location_id.usage not in ('view', 'internal') and
+            m.location_dest_id.usage in ('view', 'internal')))
+        if not received:
+            return res
+        products = received.mapped('product_id')
+        # Find pickings and relaunch reservation
+        moves_pickings = self.search([
+            ('picking_id.picking_type_subcode', '=', 'PICK'),
+            ('state', '=', 'confirmed'),
+            ('product_id', 'in', products.ids),
+            ('picking_id.printed', '!=', True)])
+        pickings = moves_pickings.mapped('picking_id')
+        _logger.debug("Products received are in backorder")
+        # unreserve moves having an operation for that product
+        # Note: (re)check availability (action_assign) does not
+        # work on added move where an operation already exists for
+        # that product. To not recompute all the quants of the
+        # picking, we delete only the pack operation to recompute.
+        # No need to perform the assignment now (new pack operation
+        # creation), it is performed later when the procurement is
+        # run.
+        operations_to_recompute = pickings.pack_operation_ids. \
+            filtered(lambda op: op.product_id in products)
+        if operations_to_recompute:
+            _logger.debug("Cleaning operations %s",
+                          operations_to_recompute.ids)
+            operations_to_recompute.mapped(
+                'linked_move_operation_ids.move_id').do_unreserve()
+        _logger.debug("Reserve corresponding moves %s", moves_pickings)
+        moves_pickings.action_assign()
+        return res
+
+    @api.multi
     @api.constrains('picking_id')
     def _check_round(self):
         if not self.mapped('picking_id.delivery_round_id'):